+++ container-engine-lib/lib/backend_interface/objects/service/files_artifact_id.go
@@ -1,4 +1,0 @@
-package service
-
-// TODO Remove this - it's not used anywhere in this package (just by dependants)
-type FilesArtifactID string